Add files via upload
here we add singleton design pattern on ScrabbleGame this would ensure that there can only be one instance of the game running at any given time. This means that if a user tries to start a new game while another game is already in progress, the Singleton pattern would prevent a new instance from being created, and instead return a reference to the existing instance.