Add files via upload
Here we add Facade Pattern to a bank app. A banking system can be a complex system with many different components and operations, such as account management, transaction processing, loan processing, and more. The Facade pattern can be used to create a simplified interface to these operations, making it easier for customers or other applications to interact with the bank's services.